refactor recursion circuit from example
